requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6 